Allow to disable the Chromium rendering pipeline
This introduces a hook and patches to Chromium which allow the embedding application to disable and enable the rendering pipeline of an embedded browser window at will. To be more exact: the pipeline is actually cut short the moment that rendered stuff would be drawn to the screen otherwise, so effectively, everything just works as it normally would (event handlers being called and such), but the screen doesn't update anymore until rendering is re-enabled.